requirements.txt
# Core Streamlit and web framework
streamlit>=1.38.0

# Build tools (prefer wheels; faster installs)
pip>=24.0
setuptools>=70.0
wheel>=0.43.0

# Data manipulation and analysis
pandas==2.2.3
numpy==2.3.2

# Data visualization
plotly==5.22.0

# Database and ORM
sqlalchemy==2.0.30
psycopg2-binary==2.9.9   # Only needed if you connect directly to PostgreSQL

# Supabase integration
supabase==2.7.4

# Environment variable management
python-dotenv==1.0.1

# Authentication and security
cryptography==42.0.5

# Additional utilities
python-dateutil==2.9.0.post0

# HTTP client pins (compat with supabase 2.7.4 and chromadb 0.5.3)
httpx==0.27.2
httpcore==1.0.5

# LangChain ecosystem
langchain==0.3.27
langchain-community==0.3.27
langchain-google-genai==2.1.9
langchain-openai==0.2.10
langchain-text-splitters==0.3.9
langsmith==0.4.5

# PDF and scraping
pypdf==4.2.0
beautifulsoup4==4.12.3
requests==2.32.3
trafilatura==1.6.3
python-frontmatter==1.0.1

# Pin mistralai for reproducible builds
mistralai==1.0.0

# Document editing for report insertion
python-docx==1.1.2

# Image analysis (Food-101)
Pillow==10.4.0
# For CPU installs of PyTorch on Windows, use the official wheels if pip resolve fails.
torch==2.5.1
torchvision==0.20.1
# Optional: broader model backbone support (not required)
# timm==1.0.9


# Voice Assistant Dependencies
vosk
sounddevice
pyttsx3
assemblyai
